+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This is the changelog for Astro version 0.2.0 released on December 6th, 2019.  F
 
 * Add `Astro.equinox/2`and `Astro.solstice/2` to calculate solstices and equinoxes for a year. From these can be derived the seasons.
 
+* Add `Astro.Time.datetime_from_julian_days/1`
+
+* Add `Astro.Time.utc_datetime_from_terrestrial_datetime/1`
+
 # Changelog for Astro version 0.1.0
 
 This is the changelog for Astro version 0.1.0 released on December 5th, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/astro/tags)